Helper function to estimate reference set(s).

The function takes an data frame with columns at least specified by obj.cols and “prob”. The reference set for each unique problem in column “prob” is then obtained by combining all approximation sets generated by all considered algorithms for the corresponding problem and filtering the non-dominated solutions.

approximateRefSets(df, obj.cols, as.df = FALSE)

Arguments

df

[data.frame]
Data frame with the required structure.

obj.cols

[character(>= 2)]
Column names of the objective functions.

as.df

[logical(1)]
Should a data.frame be returned? Default is FALSE. In this case a named list is returned.

Value

[list | data.frame] Named list of matrizes (names are the problems) or data frame with columns obj.cols and “prob”.