Initialize Pareto Archive.

A Pareto archive is usually used to store all / a part of the non-dominated points stored during a run of an multi-objective evolutionary algorithm.

initParetoArchive(control, max.size = Inf, trunc.fun = NULL)

Arguments

control

[ecr_control]
Control object.

max.size

[integer(1)]
Maximum capacity of the Pareto archive, i.e., the maximal number of non-dominated points which can be stored in the archive. Default is Inf, i.e., (theoretically) unbounded capacity.

trunc.fun

[function(archive, inds, fitness, ...)]
In case the archive is limited in capacity, i.e., max.size is not infinite, this function is called internally if an archive overflow occurs. This function expects the archive, a list of individuals inds, a matrix of fitness values (each column contains the fitness value(s) of one individual) fitness and further optional arguments ... which may be used by the internals of trunc.fun. The function must return a list with components “fitness” and “inds” which shall be the subsets of fitness and inds respectively, which should be kept by the archive.

Value

[ecr_pareto_archive]

See also