Factory method for monitor objects.

Monitor objects serve for monitoring the optimization process, e.g., printing some status messages to the console. Each monitor includes the functions before, step and after, each being a function and expecting a log log of type ecr_logger and ... as the only parameters. This way the logger has access to the entire log.

makeECRMonitor(before = NULL, step = NULL, after = NULL, ...)

Arguments

before
[function] Function called one time after initialization of the EA.

step
[function] Function applied after each iteration of the algorithm.

after
[function] Function applied after the EA terminated.

...
[any] Not used.

Value

[ecr_monitor] Monitor object.