Creates an optimization task.

An optimization task consists of the fitness/objective function, the number of objectives, the “direction” of optimization, i.e., which objectives should be minimized/maximized and the names of the objectives.

makeOptimizationTask(fun, n.objectives = NULL, minimize = NULL,
  objective.names = NULL)

Arguments

fun
[function | smoof_function] Fitness/objective function.

n.objectives
[integer(1)] Number of objectives. This must be a positive integer value unless fun is of type smoof_function.

minimize
[logical] A logical vector indicating which objectives to minimize/maximize. By default all objectives are assumed to be minimized.

objective.names
[character] Names for the objectuves. Default is NULL. In this case the names are set to y1, ..., yn with n equal to n.objectives and simply y in the single-objective case.

Value

[ecr_optimization_task]