Normalize approximations set(s).

Normalization is done by subtracting the min.value for each dimension and dividing by the difference max.value - min.value for each dimension Certain EMOA indicators require all elements to be strictly positive. Hence, an optional offset is added to each element which defaults to zero.

normalize(x, obj.cols, min.value = NULL, max.value = NULL, offset = NULL)

Arguments

x

[matrix | data.frame]
Either a numeric matrix (each column corresponds to a point) or a data.frame with columns at least obj.cols.

obj.cols

[character(>= 2)]
Column names of the objective functions.

min.value

[numeric]
Vector of minimal values of length nrow(x). Only relevant if x is a matrix. Default is the row-wise minimum of x.

max.value

[numeric]
Vector of maximal values of length nrow(x). Only relevant if x is a matrix. Default is the row-wise maximum of x.

offset

[numeric]
Numeric constant added to each normalized element. Useful to make all objectives strictly positive, e.g., located in \([1,2]\).

Value

[matrix | data.frame]

Note

In case a data.frame is passed and a “prob” column exists, normalization is performed for each unique element of the “prob” column independently (if existent).

See also