Visualize three-objective Pareto-front approximations.

Given a data frame with the results of (multiple) runs of (multiple) different three-objective optimization algorithms on (multiple) problem instances the function generates 3D scatterplots of the obtained Pareto-front approximations.

plotScatter3d(
  df,
  obj.cols = c("f1", "f2", "f3"),
  max.in.row = 4L,
  package = "scatterplot3d",
  ...
)

Arguments

df

[data.frame]
Data.frame with columns at least obj.cols, “prob” and “algorithm”.

obj.cols

[character(>= 3)]
Column names of the objective functions. Default is c("f1", "f2", "f3").

max.in.row

[integer(1)]
Maximum number of plots to be displayed side by side in a row. Default is 4.

package

[character(1L)]
Which package to use for 3d scatterplot generation? Possible choices are “scatterplot3d”, “plot3D”, “plot3Drgl” or “plotly”. Default is “scatterplot3d”.

...

[any]
Further arguments passed down to scatterplot function.

Value

Nothing

See also