Stopping conditions

Stop the EA after a fixed number of fitness function evaluations, after a predefined number of generations/iterations, a given cutoff time or if the known optimal function value is approximated (only for single-objective optimization).

stopOnEvals(max.evals = NULL)

stopOnIters(max.iter = NULL)

stopOnOptY(opt.y, eps)

stopOnMaxTime(max.time = NULL)

Arguments

max.evals

[integer(1)]
Maximal number of function evaluations. Default is Inf.

max.iter

[integer(1)]
Maximal number of iterations/generations. Default is Inf.

opt.y

[numeric(1)]
Optimal scalar fitness function value.

eps

[numeric(1)]
Stop if absolute deviation from opt.y is lower than eps.

max.time

[integer(1)]
Time limit in seconds. Default is Inf.

Value

[ecr_terminator]