Export results of statistical tests to LaTeX table(s).

Returns high-quality LaTeX-tables of the test results of statistical tests performed with function test on per-instance basis. I.e., a table is returned for each instances combining the results of different indicators.

toLatex(
  stats,
  stat.cols = NULL,
  probs = NULL,
  type = "by.instance",
  cell.formatter = NULL
)

# S3 method for list
toLatex(
  stats,
  stat.cols = NULL,
  probs = NULL,
  type = "by.instance",
  cell.formatter = NULL
)

# S3 method for data.frame
toLatex(
  stats,
  stat.cols = NULL,
  probs = NULL,
  type = "by.instance",
  cell.formatter = NULL
)

Arguments

stats

[list]
Data frame (return value of computeIndicators) or named list of list as returned by test.

stat.cols

[character]
Names of the indicators to consider. Defaults to all indicators available in stats.

probs

[character]
Filtering: vector of problem instances. This way one can restrict the size of the table(s). Defaults to all problems available in stats. Ignored if stats is a data frame.

type

[character(1)]
Type of tables. At the moment only option “by.instance” is available. I.e., a separate LaTeX-table is generated for each instance specified via probs. Ignored if stats is a data frame.

cell.formatter

[function(cell, ...)]
Function which is used to format table cells. This function is applied to each table cell and may be used to customize the output. Default is niceCellFormater. Ignored if stats is a data frame.

Value

[list] Named list of strings (LaTeX tables). Names correspond to the selected problem instances in probs.

See also