Update the log.

This function modifies the log in-place, i.e., without making copies.

updateLogger(log, population, fitness = NULL, n.evals, ...)

Arguments

log
[ecr_logger] The log generated by initLogger.

population
[list] List of individuals.

fitness
[matrix] Optional matrix of fitness values (each column contains the fitness value(s) for one individual) of population. If no matrix is passed and the log shall store information of the fitness, each individual needs to have an attribute fitness.

n.evals
[integer(1)] Number of fitness function evaluations performed in the last generation.

...
[any] Furhter arguments. Not used at the moment.

See also

Other logging: getPopulations, getStatistics, initLogger