Result object.

S3 object returned by ecr containing the best found parameter setting and value in the single-objective case and the Pareto-front/-set in case of a multi-objective optimization problem. Moreover a set of further information, e.g., reason of termination, the control object etc. are returned.

The single objective result object contains the following fields:

task

The ecr_optimization_task.

best.x

Overall best parameter setting.

best.y

Overall best objective value.

log

Logger object.

last.population

Last population.

last.fitness

Numeric vector of fitness values of the last population.

message

Character string describing the reason of termination.

In case of a solved multi-objective function the result object contains the following fields:

task

The ecr_optimization_task.

log

Logger object.

pareto.idx

Indizes of the non-dominated solutions in the last population.

pareto.front

(n x d) matrix of the approximated non-dominated front where n is the number of non-dominated points and d is the number of objectives.

pareto.set

Matrix of decision space values resulting with objective values given in pareto.front.

last.population

Last population.

message

Character string describing the reason of termination.