Computation of the unary epsilon-indicator.

Functions for the computation of unary and binary measures which are useful for the evaluation of the performace of EMOAs. See the references section for literature on these indicators.

Given a set of points points, emoaIndEps computes the unary epsilon-indicator provided a set of reference points ref.points.

The emoaIndHV function computes the hypervolume indicator Hyp(X, R, r). Given a set of point X (points), another set of reference points R (ref.points) (which maybe the true Pareto front) and a reference point r (ref.point) it is defined as Hyp(X, R, r) = HV(R, r) - HV(X, r).

emoaIndEps(points, ref.points)

emoaIndHV(points, ref.points, ref.point = NULL)

emoaIndR1(points, ref.points, ideal.point = NULL, nadir.point = NULL,
  lambda = NULL, utility = "tschebycheff")

emoaIndR2(points, ref.points, ideal.point = NULL, nadir.point = NULL,
  lambda = NULL, utility = "tschebycheff")

emoaIndR3(points, ref.points, ideal.point = NULL, nadir.point = NULL,
  lambda = NULL, utility = "tschebycheff")

Arguments

points
[matrix] Matrix of points.

ref.points
[matrix] Set of reference points.

ref.point
[numeric] A single reference point used, e.g., for the computation of the hypervolume indicator via emoaIndHV. If NULL the nadir point of the union of the points and ref.points is used.

ideal.point
[numeric] The utopia point of the true Pareto front, i.e., each component of the point contains the best value if the other objectives are neglected.

nadir.point
[numeric] Nadir point of the true Pareto front.

lambda
[integer(1)] Number of weight vectors to use in estimating the utility function.

utility
[character(1)] Name of the utility function to use. Must be one of “weightedsum”, “tschebycheff” or “augmented tschbycheff”.

Value

[numeric(1)] Epsilon indicator.