Determine which points of a set are (non)dominated.

Given a matrix with one point per column which.dominated returns the column numbers of the dominated points and which.nondominated the column numbers of the nondominated points. Function isMaximallyDominated returns a logical vector with TRUE for each point which is located on the last non-domination level.

which.dominated(x)

which.nondominated(x)

isMaximallyDominated(x)

Arguments

x

[matrix]
Numeric (n x d) matrix where n is the number of points and d is the number of objectives.

Value

[integer]

Examples

  data(mtcars)
  # assume we want to maximize horsepower and minimize gas consumption
  cars = mtcars[, c("mpg", "hp")]
  cars$hp = -cars$hp
  idxs = which.nondominated(as.matrix(cars))
  print(mtcars[idxs, ])
#>               mpg cyl disp  hp drat    wt  qsec vs am gear carb
#> Mazda RX4 Wag  21   6  160 110  3.9 2.875 17.02  0  1    4    4